Update and-where-or-where.md - Round #1
I reworded a lot to (hopefully) make things flow a little better. I'm also not sure if I managed to catch all parts of the code that needed some parenthesis, arrows, or other things. The AI edit misses some of that, and I'm also not sure if "andWhere" and "orWhere" need the "()" at the end every time it shows up. I added it in some cases, but not other.