Update joins.md - Round #1
I reworded a ton here because, strangely, pretty much one whole paragraph was just missing, and the beginning of the audio started out different from the text. So much so, that for a moment, I thought I might have the wrong video file pulled up. I think I've accounted for everything that was missing, but definitely double check to make sure nothing important was left out.