Update select-sum.md - Round #1
I re-wrote A LOT of this (partly because several lines were missing, but also to improve the flow a little) so definitely double check me. I also couldn't remember how to format those "basically, this is saying [insert information here]" bits like you'll see on line 11, so I apologize in advance if that's not quite right. There's a lot to remember about these specific moments and it's all dancing around in my brain.