Update select-specific-fields.md - Round #1
Around line 23, you were basically like "there's no actual reason to do that" and I did a tiny facepalm because I was like "why are we doing it, then". XD But you explained the reason a little better further along. Maybe there's a better way to word that so people going through the tutorial don't get to that part and go "why tho" [insert that little priest meme here]. Anyway, this one was pretty on par with the last script. I rewrote a bit for flow, so just make sure that all checks out.