Update reusing-queries-query-builder.md - Round #1
I'm not super sure if I had to put "QueryBuilder" in fancy ticks as much as I did, or if I was supposed to say "$qb" in some places instead of spelling it out. Also, fun fact: If you Google search "PHP criteria" you get a bunch of information about "patient hospitalization programs". XD I only wanted to check and see if that needed to be capitalized or not.